Major update to git. Application entering finalazing form.

+++ COMP-FISK/Controllers/RacuniController.cs
@@ -4,6 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
+using Tulpep.NotificationWindow;
 
 namespace COMP_FISK.Controllers
 {
@@ -24,13 +26,23 @@ namespace COMP_FISK.Controllers
             var putanjaFROM_FP = @"C:\fiskcomp\exch\lnk\FROM_FP";
             TextWriter odgovor = new StreamWriter(putanjaFROM_FP + @"\" + Path.GetFileNameWithoutExtension(nazivFajla) + ".OK");
             odgovor.WriteLine(brojRacuna);
-            odgovor.Flush();
+            odgovor.Close();
             PremjestiDbf(@"C:\fiskcomp\exch\lnk\TO_FP\" + nazivFajla, @"C:\fiskcomp\exch\lnk\comp\" + nazivFajla);
+            
         }
 
         public static void PremjestiDbf(string putanjaFajla, string destinacija)
         {
-            File.Move(putanjaFajla, destinacija);
+            var fajlBezExtenzije = Path.GetFileNameWithoutExtension(putanjaFajla);
+            var fullPutanja = @"C:\fiskcomp\exch\lnk\Neodstampani\" + fajlBezExtenzije + ".dbf";
+
+            if (File.Exists(fullPutanja))
+            {
+                File.Delete(fullPutanja);
+                File.Move(putanjaFajla, destinacija);
+            }
+            else
+                File.Move(putanjaFajla, destinacija);
         }
 
         public static void UpisiOdgovorERR(string greska, string nazivFajla)
@@ -38,7 +50,7 @@ namespace COMP_FISK.Controllers
             var putanjaFROM_FP = @"C:\fiskcomp\exch\lnk\FROM_FP";
             TextWriter odgovor = new StreamWriter(putanjaFROM_FP + @"\" + Path.GetFileNameWithoutExtension(nazivFajla) + ".ERR");
             odgovor.Write(greska);
-            odgovor.Flush();
+            odgovor.Close();
             PremjestiDbf(@"C:\fiskcomp\exch\lnk\TO_FP\" + nazivFajla, @"C:\fiskcomp\exch\lnk\Neodstampani\" + nazivFajla);
         }
     }
